backend/requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
attrs
bcrypt
certifi
cffi
charset-normalizer
click
colorama
cryptography
databases
datasets
dill
dnspython
ecdsa
email_validator
fastapi
fastapi-users
filelock
frozenlist
fsspec
greenlet
h11
huggingface-hub
idna
Jinja2
makefun
Mako
MarkupSafe
mpmath
multidict
multiprocess
networkx
numpy
packaging
pandas
passlib
pillow
propcache
psutil
pwdlib
pyarrow
pyasn1
pycparser
pydantic
pydantic_core
PyJWT
PyMySQL
python-dateutil
python-dotenv
python-jose
python-multipart
pytz
PyYAML
regex
requests
rsa
safetensors
sentencepiece
six
sniffio
SQLAlchemy
starlette
sympy
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
tzdata
urllib3
uvicorn
xxhash
yarl